% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_csv.R
\name{create_manifest}
\alias{create_manifest}
\title{Create the manifest file}
\usage{
create_manifest(
  manifest.version = "1.0",
  oneroster.version = "1.1",
  path = "~/",
  upload_type = "bulk",
  source.systemName = "package",
  source.systemCode = "NA"
)
}
\arguments{
\item{manifest.version}{the version of the manifest Must initially be 1.0}

\item{oneroster.version}{Version number of one roster (currently must be 1.1)}

\item{path}{path to folder with to create manifest for}

\item{upload_type}{A upload must either be of type 'bulk' or 'delta'}

\item{source.systemName}{the name of the system producing the files}

\item{source.systemCode}{ID code of the system producing the files}
}
\value{
a tibble
}
\description{
Create the manifest file
}
