% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_csv.R
\name{manifest_file_check}
\alias{manifest_file_check}
\title{Check Files for Manifest}
\usage{
manifest_file_check(path, file_name, upload_type)
}
\arguments{
\item{path}{path to file}

\item{file_name}{name of file to check for}

\item{upload_type}{the type of file upload}
}
\value{
the value for the manifest.
}
\description{
An internal helper function to check files for creation of the manifest
}
\keyword{internal}
