% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneRoster_package.R
\docType{package}
\name{oneRoster-Package}
\alias{oneRoster-Package}
\title{A R Package for One Roster Preparation}
\description{
One Roster is a standard developed by LMS Global allowing for ease of transfer of student data between
many educational systems. Creating a system of CSV compatible tables for ease of creation with Learning Management Systems (LMS),
Student Information Systems (SIS), and other learning platforms.

Data files in CSV format are exchanged using a zip file and must be at the root level.
}
\details{
The CSV files supported are as folows\tabular{lll}{
   File Name \tab Req \tab Description \cr
   manifest.csv \tab Yes \tab A control file. The manifest contains the version (set as 1.1) and the list of files that are supplied in this data set. A new file added in V1.1. \cr
   academicSessions.csv \tab No \tab A data file. The 'academic sessions' data model content. Data model revised in v1.1. \cr
   categories.csv \tab No \tab A data file. The definition for a set of lineItems. A new file added in V1.1. \cr
   classResources.csv \tab No \tab A data file. The set of resources assigned to classes. A new file added in V1.1. \cr
   classes.csv \tab No \tab A data file. The 'classes' data model content. Data model revised in v1.1. \cr
   courseResources.csv \tab No \tab A data file. The set of resources assigned to courses. A new file added in V1.1. \cr
   courses.csv \tab No \tab A data file. The 'courses' data model content. Data model revised in v1.1. \cr
   demographics.csv \tab No \tab A data file. The 'demographics' data model content. Data model revised in v1.1. \cr
   enrollments.csv \tab No \tab A data file. The 'enrollments' data model content. Data model revised in v1.1. \cr
   lineItems.csv \tab No \tab A data file. The set of lineItems. A new file added in V1.1. \cr
   orgs.csv \tab No \tab A data file. The' orgs' data model content. \cr
   results.csv \tab No \tab A data file. The set of results. A new file added in V1.1. \cr
   resources.csv \tab No \tab A data file. The 'resources' data model content. A new file added in V1.1. \cr
   users.csv \tab No \tab A data file. The 'users' data model content. Data model revised in v1.1. \cr
}


A record in a file can be in 1 of 4 sates.
\itemize{
\item \strong{No Record}: The record has not been created or has been removed and cannot be recovered
\item \strong{Record[-]}: The record has been created and assigned a sourceID but is not yet active
\item \strong{Record[Active]}: The record status has been set to 'Active'
\item \strong{record[ToBeDeleted]}: The record status has been set to 'tobedeleted'
}

For Each state there are four types of transition events
\itemize{
\item \strong{B[-]}: A bulk transfer has occurred and an existing record does not appear in the files.
\item \strong{B[S]}: A bulk transfer has occurred and the record occurs in the new file set.
\item \strong{D[A]}: A delta transfer has occurred and the record status is set as 'active'.
\item \strong{d[D]}: A delta transfer has occurred and the record status is set as 'tobedeleted'
}
}
