% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checks.R
\docType{data}
\name{oneRoster_file_columns}
\alias{oneRoster_file_columns}
\title{Files and their column names}
\format{An object of class \code{list} of length 13.}
\usage{
oneRoster_file_columns
}
\value{
a list of files with associated column names
}
\description{
A list of all one roster files and their required case sensitive file names.
}
\keyword{internal}
